requirements.txt
python-dotenv
PyAudio
SpeechRecognition
pandas